ras-frontstage/requirements.txt
Flask==0.12
pystache==0.5.4
requests==2.11.1
flask-cors==3.0.0
Flask-SQLAlchemy==2.1
Flask-Migrate==1.8.0
Flask-WTF==0.14.2
python-jose==1.2.0
psycopg2==2.6.1
Flask-OpenID==1.2.5
Flask-OAuth==0.12
requests-oauthlib==0.8.0